refactor(composer_rpc): split l2_to_l1.rs into directory module (3.7h)
Split l2_to_l1.rs (3301 LOC) into:
- mod.rs (752 LOC): HTTP handler + detection pipeline — UNDER 800 TARGET
- process.rs (1383 LOC): post-discovery orchestration + queuing
- enrichment.rs (1184 LOC): L2 trace enrichment + simulation helpers

Public API unchanged. Re-exports preserve cross-module callers.

616/616 tests, clippy clean.
